+++ src/Rest/RestApi.php
@@ -17,6 +17,7 @@ use WebcafeinaReservas\Rest\Controllers\SpacesController;
 use WebcafeinaReservas\Rest\Controllers\Admin\AdminBookingsController;
 use WebcafeinaReservas\Rest\Controllers\Admin\AdminBookingsExportController;
 use WebcafeinaReservas\Rest\Controllers\Admin\AdminCalendarController;
+use WebcafeinaReservas\Rest\Controllers\Admin\AdminHealthController;
 use WebcafeinaReservas\Rest\Controllers\Admin\AdminPdfTemplatesController;
 use WebcafeinaReservas\Rest\Controllers\Admin\AdminSettingsController;
 use WebcafeinaReservas\Rest\Controllers\Admin\AdminStatsController;
@@ -47,6 +48,7 @@ final class RestApi {
         ( new AdminPdfTemplatesController() )->register();
         ( new AdminBookingsExportController() )->register();
         ( new AdminCalendarController() )->register();
+        ( new AdminHealthController() )->register();
     }
 
     /**